# [quant][fx][perf] improve runtime of prepare step for large models