Stack from [ghstack](https://github.com/ezyang/ghstack):
* **#61132**

Summary:
For large models, the insert_observers_for_model function was taking a long time, especially for the case where not all the nodes are being quantized

For example for a model with 21000 nodes of which only ~50 are being quantized the breakdown of prepare_fx vs convert fx was

prepare_fx 979 seconds
convert_fx 9 seconds

The main reason was because we were doing some unnecessary computation for all nodes in this function, this PR just moves them to where they are actually used

After this PR
prepare_fx 26 seconds
convert_fx 9 seconds

Test Plan:
Existing tests

Reviewers:

Subscribers:

Tasks:

Tags:

Differential Revision: [D29522303](https://our.internmc.facebook.com/intern/diff/D29522303)